vrs-users/requirements.txt
requests==2.25.1
fastapi==0.110.0
python-dotenv==1.0.1
uvicorn==0.27.1
pymongo==4.6.2
passlib==1.7.4
pyJWT==2.3.0